+++ test/entity_serializer_test.dart
@@ -1,15 +1,16 @@
-import 'package:test/test.dart';
+//import 'package:test/test.dart';
 
 void main() {
-  group('A group of tests', () {
-    //final awesome = Awesome();
-
-    setUp(() {
-      // Additional setup goes here.
-    });
-
-    test('First Test', () {
-      //expect(awesome.isAwesome, isTrue);
-    });
-  });
+  //Nothing yet...
+  // group('A group of tests', () {
+  //   //final awesome = Awesome();
+  //
+  //   setUp(() {
+  //     // Additional setup goes here.
+  //   });
+  //
+  //   test('First Test', () {
+  //     //expect(awesome.isAwesome, isTrue);
+  //   });
+  // });
 }