+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Feature: type now can be given in xml node name, no need to use field node
+- Feature: add comment to generated field (attribute comment)
+- Improvement: More helper functions (duration)
+
 ## 0.0.1
 
 - Initial version.